+++ app/controllers/sessions_controller.rb
@@ -5,12 +5,14 @@ class SessionsController < ApplicationController
   def new; end
 
   def create
-    user = User.find_by(email: params[:session][:email].downcase)
+    @user = User.find_by(email: params[:session][:email].downcase)
 
-    if user&.authenticate(params[:session][:password])
+    if @user&.authenticate(params[:session][:password])
       reset_session
-      log_in user
-      redirect_to user
+      params[:session][:remember_me] == '1' ? remember(@user) : forget(@user)
+
+      log_in @user
+      redirect_to @user
     else
       flash.now[:danger] = 'Invalid email/password combination'
       render 'new', status: :unprocessable_entity
@@ -18,7 +20,7 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-    log_out
+    log_out if logged_in?
     redirect_to root_url, status: :see_other
   end
 end