+++ app/models/user.rb
@@ -2,6 +2,8 @@
 
 # The user model
 class User < ApplicationRecord
+  attr_accessor :remember_token
+
   VALID_EMAIL_REGEX = /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i
 
   before_save { email.downcase! }
@@ -13,8 +15,40 @@ class User < ApplicationRecord
 
   has_secure_password
 
+  # Returns the has digest of the given string
   def User.digest(string)
     cost = ActiveModel::SecurePassword.min_cost ? BCrypt::Engine::MIN_COST : BCrypt::Engine.cost
     BCrypt::Password.create string, cost: cost
+  end
+
+  # Returns a random token
+  def User.new_token
+    SecureRandom.urlsafe_base64
+  end
+
+  # Remembers a token in the database for use in a persistent session
+  def remember
+    self.remember_token = User.new_token
+    update_attribute(:remember_digest, User.digest(remember_token))
+    remember_digest
+  end
+
+
+  # remembers a session token to prevent session hijacking.
+  # We reuse the remember digest for convenience
+  def session_token
+    remember_digest || remember
+  end
+
+  # Returns true if the given token matches the digest
+  def authenticated?(remember_token)
+    return false if remember_digest.nil?
+    
+    BCrypt::Password.new(remember_digest).is_password? remember_token
+  end
+
+  # forgets a user
+  def forget
+    update_attribute(:remember_digest, nil)
   end
 end