+++ test/integration/users_login_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "test_helper"
+require 'test_helper'
 
 class UsersLogin < ActionDispatch::IntegrationTest
   def setup
@@ -66,5 +66,45 @@ class LogoutTest < Logout
     assert_select 'a[href=?]', login_path
     assert_select 'a[href=?]', logout_path, count: 0
     assert_select 'a[href=?]', user_path(@user), count: 0
+  end
+
+  test 'login with valid information followed by logout' do
+    post login_path, params: { session: { email: @user.email, password: 'password' } }
+    assert is_logged_in?
+    assert_redirected_to @user
+    follow_redirect!
+
+    assert_template 'users/show'
+    assert_select 'a[href=?]', login_path, count: 0
+    assert_select 'a[href=?]', logout_path
+    assert_select 'a[href=?]', user_path(@user)
+
+    delete logout_path
+    assert_response :see_other
+    assert_not is_logged_in?
+    assert_redirected_to root_url
+    # Simulate user cliging logout in a second window
+    delete logout_path
+    follow_redirect!
+    assert_select 'a[href=?]', login_path
+    assert_select 'a[href=?]', logout_path, count: 0
+    assert_select 'a[href=?]', user_path(@user), count: 0
+  end
+
+  test 'login with remembering' do
+    # Log in to set the cookie
+    log_in_as @user, remember_me: '1'
+
+    assert_not cookies[:remember_token].blank?
+    assert_equal cookies[:remember_token], assigns(:user).remember_token
+  end
+
+  test 'login without remembering' do
+    # Login to set cookie
+    log_in_as @user, remember_me: '1'
+
+    # login again and verify that the cookie is deleted
+    log_in_as @user, remember_me: '0'
+    assert cookies[:remember_token].blank?
   end
 end