+++ CHANGELOG.md
@@ -5,13 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-About to start [Chapter 9](https://www.learnenough.com/ruby-on-rails-7th-edition-tutorial/advanced_login)
+About to start [Chapter 9.3.2 Testing the remember branch](https://www.learnenough.com/ruby-on-rails-7th-edition-tutorial/advanced_login#sec-testing_the_remember_branch)
 
 ## TODO
 
 - Finish configuring rails-erd after first models configured.
 
 ## [Unreleased]
+
+- added remember token
 
 ## v0.0.7 2022-05-01